#3601 use scale parameter
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@41632 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ web/src/main/java/org/geogebra/web/html5/util/ArticleElement.java
@@ -483,11 +483,11 @@ public final class ArticleElement extends Element {
 		return ret == null ? "" : ret;
 	}
 
-	public double getDataParamScale() {
+	public float getDataParamScale() {
 		String scale = this.getAttribute("data-param-scale");
-		double ret = 1;
+		float ret = 1;
 		try {
-			ret = Double.valueOf(scale);
+			ret = Float.valueOf(scale);
 		} catch (Throwable t) {
 			Log.warn("Invalid scale");
 		}
